+++ rich/__main__.py
@@ -86,7 +86,7 @@ def make_test_card() -> Table:
         return table
 
     table.add_row(
-        "Chinese Japanese & Korean support",
+        "CJK support",
         Panel("该库支持中文，日文和韩文文本！", expand=False, style="red", box=box.DOUBLE_EDGE,),
     )