+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.7] - 2020-03-31
+
+### Fixed
+
+- Broken wrapping of long lines
+- Fixed wrapping in Syntax
+
+### Changed
+
+- Added word_wrap option to Syntax, which defaults to False.
+- Added word_wrap option to Traceback.
+
+## [0.8.6] - 2020-03-29
+
+### Added
+
+- Experimental Jupyter notebook support: from rich.jupyter import print
+
 ## [0.8.5] - 2020-03-29
 
 ### Changed